+++ R/linalg.R
@@ -28,6 +28,7 @@ solve.mlx <- function(a, b = NULL, ..., device = NULL) {
   if (is.null(b)) {
     a <- mlx_cast(a, dtype = target_dtype)
     ptr <- cpp_mlx_solve(a$ptr, NULL, target_dtype)
+    dimnames <- dimnames_matrix_inverse(a)
   } else {
     if (!is_mlx(b)) {
       b <- as_mlx(b, dtype = target_dtype)
@@ -38,9 +39,10 @@ solve.mlx <- function(a, b = NULL, ..., device = NULL) {
     }
 
     ptr <- cpp_mlx_solve(a$ptr, b$ptr, target_dtype)
+    dimnames <- dimnames_solve(a, b)
   }
 
-  new_mlx(ptr)
+  new_mlx(ptr, dimnames = dimnames)
 }
 
 #' Kronecker product dispatcher